+++ Models/UnitCircleCoordinates.cs
@@ -4,7 +4,7 @@ namespace GraphicsAlgorithmsApp.Models
 {
     public static class UnitCircleCoordinates
     {
-        public static (double, double) ScreenToUnitCircle(int screenX, int screenY, int centerX, int centerY, double scale)
+        public static (double, double) ScreenToUnitCircle(double screenX, double screenY, double centerX, double centerY, double scale)
         {
             double unitX = (screenX - centerX) / scale;
             double unitY = -(screenY - centerY) / scale;
@@ -12,10 +12,10 @@ namespace GraphicsAlgorithmsApp.Models
             return (unitX, unitY);
         }
         
-        public static (int, int) UnitCircleToScreen(double unitX, double unitY, int centerX, int centerY, double scale)
+        public static (double, double) UnitCircleToScreen(double unitX, double unitY, double centerX, double centerY, double scale)
         {
-            int screenX = (int)(centerX + unitX * scale);
-            int screenY = (int)(centerY - unitY * scale);
+            double screenX = (centerX + unitX * scale);
+            double screenY = (centerY - unitY * scale);
             
             return (screenX, screenY);
         }
